Fixing the branch distance calculation (true/false)
To determine whether the branch to cover is true or false branch, we need to look at the attribute "type" and not the "locatiionId". This commit fixes the issues. I have tested the new verson against the entire benchmark. I have tested this fix against the benchmark and it works correclty.