Use mock to cleanup setup for tests
The UI is used as a logging API and used in many classes.
Its setup requires a series of multiple calls setting global variables

Replace this setup by a simple (sinon) mock of the UI.

This also makes it mandatory to explicitly set properties that steer
the algorithms (making the tests self documenting and less fragile).

Setting a property was only needed for one test case:
Testing preference sorting in MOSA.