feat: modular plugin system (#171)
With the current architecture, it would be difficult for researchers and developers to propose other solutions (e.g., search algorithms, search operators, and research improvements) without maintaining forks and us having to support their contributions. With this new modular architecture, we provide developers the opportunity to add functionality through plugins that they can easily add and support. This should make it easier for researchers to propose new changes.

Functionality that is critical to the core working of the framework should still be included in the main repo. The plugins mostly serve as one-off implementations and experimentation trials.

Co-authored-by: Mitchell Olsthoorn <mitchell.olsthoorn@outlook.com>