feat: first class plugins architecture (#234)
Our idea behind the SynTest-Framework is to create one suite of tools that help developers maintain a high standard of code quality. To accomplish this, we plan to incorporate many different tools over time. To make this possible, we have to change the architecture to make plugins a first-class citizen.