ci: add main branch workflows and set fetch-depth to 0 (#244)
* ci: add main branch workflows

Previously, we only had build and quality workflows for PRs. To allow us
to show build and quality status badges on the README, we need to also
run these workflows on the main branch.

* ci: set fetch-depth to 0

The default fetch-depth of 1 only gets the latest commit without any tags or history.
For the lerna version/publish workflows it requires the git history.