feat: improve modularity of secondary objectives (#248)
Previously, we hardcoded the secondary objectives in the objective managers. This does not allow users to change the secondary objective or develop new ones.

All these modularity improvements make the search algorithms less and less like their original versions in the paper. This PR solves this by adding presets: a way to allow a group of config options to be set to a specific value.

---------

Co-authored-by: Dimitri Stallenberg <tweet.dimitri@gmail.com>