feat: generalize cfg package (#256)
The CFG package used a simplified model of the concept related to the control flow graph. The CFG also contained all nodes for an entire file instead of per executable method. This PR refactors the package to have a more generalized structure. Additionally, this PR adds an edge contraction algorithm to compress the graph.