fix: metric file writer (#302)
The metric file writer was not writing the metrics to the disk. This had to do with the event 'exit'. We put a listener on an event called 'exit', which is the same as the system-level exit event. Performing async actions on the exit event is impossible as it will never reach the next async event loop. So the code was never executed. We solved this by using the pre-exit event.