feat: experimental optimized objective manager
The SimpleObjectiveManager optimizes all objectives regardless if they
are covered or not. The UncoveredObjectiveManager and the
StructuralObjectiveManager only optimize uncovered objectives. This is
done to speed up the evaluation process.

This new objective manager adds a structural objective manager that also
optimizes the covered objectives. Giving the secondary objectives
a chance to improve the already covered objectives.