fix(search): tournament selection
In the old version, when a selected solution had a large crowding
distance than the current winner, it was prefered over the winner. This
should not happen.

The comparison should first be based on the ranking and only than on the
crowding distance.

The winner is the solution with the best (smaller) non-dominance rank.
At the same level or ranking, the winner is the solution with the best
(largest) crowding distance.