fix: NPM clean script (#140)
The clean script in the NPM config makes calls to other NPM scripts. However, it didn't prepend the calls with 'npm run' making it fail. Additionally, a call to NX was made after the node_modules directory is deleted, resulting in the script not finding the binary.

This PR also updates the NPM lockfile to version 3 corresponding to the latest LTS Node.JS version