feat: improve archive with usage tracking
Previously the archive mainly only kept track of the objectives.
However, since the search process is focussed on test case generation,
it makes sense to keep track of the unique encodings stored in the
archive as well.

Additionlly, this allows for easily extracting the encoding when parsing
the final test suite and showing which objectives an encoding was
selected for.