feat: improve objective managers (#327)
* feat: improve archive with usage tracking

Previously the archive mainly only kept track of the objectives.
However, since the search process is focused on test case generation,
it makes sense to keep track of the unique encodings stored in the
archive as well.

Additionally, this allows for easily extracting the encoding when parsing
the final test suite and showing which objectives an encoding was
selected for.

* feat: improve updating mechanism

The caller can now decide if encodings are not selected for any objective
should be kept in the archive or be deleted

* feat: add shallow distance calculation for the objectives

Shallow distances allow for speeding the evaluation of already covered
objectives while still searching for better solutions than the initial
encoding that covers an objective.

* fix: population-based search algorithms

We now distinguish between two types of objective managers: (1) Archive-based (which is what we used before) and (2) Population-based. This is done because certain search algorithms were optimized beyond what they should have been. Therefore, making comparison incorrect.