feat: integer statements (#150)
Currently, we only have a statement for the JavaScript Numeric type, which encompasses integers and decimals. However, most problems will require integers and not doubles. Using the numeric type makes it much more challenging to satisfy certain branches.

This PR adds an integer type next to the default numeric type so that the search algorithm can figure out which type works better.