feat: make seeded prng global (#348)
Currently, the seeded PRNG is only used internally by us. However, the
library we use for this has an option to overwrite Math.random(). This
would allow us to catch any random call that we might have missed in our
code but also calls to Math.random() in our dependencies.

We need to be careful though that this does not have unwanted side
affects on our dependencies, but this should not be the case.