feat: remove evaluation argument from budgets (#341)
The evaluation argument is never used in the budgets and therefore can be removed. This allows us to clean up the objective manager a bit.

Originally, we wanted to remove all arguments to the budgets. However, as it turns out the stagnation budget requires the search algorithm to determine if any progress was made