feat: improved testcase readability (#176)
* Created assertion data object such that we do not have to rely on logs to file anymore
* Used the context builder to create better variable names
* Enable assertions again
* Replace try catch for exceptions by proper "rejectedWith" function from chai

I also found a bug:
The require function caches contents of files. So when we require a "constant object" from a file multiple times we are actually using the same instance. During the search process this is no problem since all tests are ran in isolation. However, the final test suite puts all the tests in the same file so the results where not consistent.

I fixed this by removing cached files in the beforeEach method of mocha.