chore: rename core to framework (#382)
Based on #380, we decided to start the process of renaming this repo to framework from the current name core. The main reasons for these are:

The name core is not very suitable anymore as the core now contains not only common libraries for the other repositories but also the new modular extendable feature set. This allows people to build tools like our other repos on top of the framework.

It is confusing what the framework actually refers to (1) all repos combined, or (2) the modular extendable ecosystem contained in this repo. We think it should be the second, therefore, we have decided to rename the repo to framework.

Additionally:
As the plugins were renamed, the test workflows weren't working anymore. Given that the workspaces were managed manually before, I decided to automate the process of collecting the workspace paths. It took some tries to get the new CI workflow to work.

All directories on the second level with a package.json (e.g., //package.json) are selected as a workspace. This caused some crashes as not all workspaces produce coverage files. Therefore, I have added an additional step that checks if the coverage file is present before collecting the coverage and sending it through.

The workflow will only fail when there is a coverage file but the coverage collection step cannot process it.